chore: cleanup targets
