+++ src/index.ts
@@ -2,6 +2,5 @@ export * from './constants'
 export * from './hooks'
 export * from './reconciler'
 export * from './renderer'
-export * from './utils'
-export * from './shared'
 export * from './types'
+export * from './utils'
