--- src/shared/hooks.ts
@@ -1,6 +1,0 @@
-import * as React from 'react'
-
-/**
- * An SSR-friendly useLayoutEffect.
- */
-export const useIsomorphicLayoutEffect = typeof window !== 'undefined' ? React.useLayoutEffect : React.useEffect
