--- src/test/index.ts
@@ -1,2 +1,0 @@
-export * from './render'
-export * from './WebGLRenderingContext'
